requirements.txt
cassandra-driver
faker